Re-organise code to better reflect interview expectations
1. Move chart.js & reading.js into utils/ for organisation
2. Move reading manipulation utilities from chart.js -> reading.js. chart.js should only be concerned with rendering the chart.
3. Remove the default chart rendering export. This needs to be inline at the usage so that candidate can work through possible refactors on this
4. Add extra containerId parameter to renderChart to correct the coupling issue and keeping code DRY.
