+++ components/App.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Sidebar } from "./Sidebar.jsx";
 import { EnergyConsumption } from "./EnergyConsumption.jsx";
-import { getReadings } from "../reading";
+import { getReadings } from "../utils/reading";
 
 export const App = () => {
   const [readings, setReadings] = useState();
